Check if runtime array check is needed for Unsafe.setMemory() on PPC
For codegen of Unsafe.setMemory() on power, check if object type is known
at compile time. If so, calculate and pass in final destination address to
setMemoryEvaluator() by adjusting dest and destOffset if necessary
(i.e.: offheap enabled and object is array) and then adding them. Otherwise,
pass in base address and offset separately so that an arrayCHK and argument
adjustments can be generated at runtime.

Signed-off-by: midronij <jackie.midroni@ibm.com>