Bug fix (https://github.com/eclipse-openj9/openj9/issues/19947)
A crash was occuring in Unsafe.get/put() case (6) when trying to modify
the CFG to include the newly generated directAccessBlock. To avoid this,
we can simply append the directAccessTreeTop onto the previous block
(beforeCallBlock), which elimiates the need to update the CFG, since no
new blocks are added. Additionally, without this commit, these two blocks
(in addition to the one after the directAccessBlock, joinBlock), would
get merged during block simplication anyways: so this fix eliminates the
need to perform that optimization later down the line.

Signed-off-by: midronij <jackie.midroni@ibm.com>