Use baseObj for Unsafe.CAS CardMarking for OffHeap on AArch64
When evaluating Unsafe.CAS while running with Balanced GC and OffHeap
enabled, if the object child is the dataAddrPointer load, pass the
baseObj to VMCardCheckEvaluator for correct card marking.

The dstReg and temp2Reg in VMCardCheckEvaluator can share a reg,
adding the argument clobberDstReg to indicate if dstReg can be used
as the temp2Reg.

Not allocating temp2Reg and adding a dep for baseObjReg uses the same
number of registers.

Signed-off-by: Abdulrahman Alattas <rmnattas@gmail.com>